requirements.txt
bcrypt==3.1.3
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.0.4
Flask-SQLAlchemy==2.2
requests==2.4.2
psycopg2==2.7.3.1
oauth2client==3.0.0
google-api-python-client==1.6.4
google-auth==1.2.0
google-auth-oauthlib==0.2.0
Flask-Script==2.0.6
celery==4.1.0
redis==2.10.6